ltpl_test/requirements.txt
asgiref==3.4.1
Django==3.2.13
django-restframework==0.0.1
djangorestframework==3.13.1
numpy==1.19.5
opencv-python==3.3.0.9
Pillow==8.4.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
